Replaced system time with CARLA game time
This commit replaces the use of the system time in the ScenarioManager
and different scenario behaviors with the use of the CARLA game time.

Therefore, a new Python module was added, timer.py, which includes
access to the CARLA game time, and provide a new timeout behavior using
the CARLA game time.

Other changes:
- KeepVelocity behavior is now time independent. If a duration is
desired, a parallel maneuver has to be used. The example scenario is
adjusted accordingly.
- Scenario updates are now performed on every CARLA tick
- Added debug flag to ScenarioManager to minimize output.

Change-Id: I6327a7e82ac9f71e8d270d798af7857f7cfcced7