Make use of hero vehicle identification
This commit introduces the usage of a hero vehicle identification, that
was integrated into CARLA with commit bc5a3ed.

As a result, the hero (ego) vehicle has to be only defined in the
scenario, and all other clients can directly access it, without knowing
the id or type of vehicle.

Change-Id: I6d93434d2f13626a17fb4ccc7f462ae8774d4497