Reworked scenario execution to run a sequence of scenarios
This is a commit to prepare the CARLA challenge, where a sequence of
scenarios will be executed. This sequence will be created on the fly,
i.e. while the ego vehicle drives around.

Changes:
- Added configuration files (xml-format) for each scenario to have a
fixed configuration for testing
- Added configuration parser for the new config files
- Reworked scenario execution: All actors are now provided via the
scenario constructor and are no longer handled internally inside the
scenario. Instead these are maintained outside.
- Scenarios that work already:
* FollowLeadingVehicle

ToDos:
- Update scenarios:
* control_loss
* object_crash_intersection
* no_signal_junction_crossing
* object_crash_vehicle
* opposite_vehicle_taking_priority

- Update documentation:
* Create new scenario
* Changelog
* Configuration parser

Change-Id: If70c2a972d7c30d46d080bb08ec87020e5c03d78