Added CarlaActorPool class to share actors among scenarios
With the commit the creation of the actors for scenarios has changed.
While the ego vehicle is still created outside (by the scenario_runner),
all other actors are created by the scenario(s).

To be able to hand over actors from one scenario to the scenario_runner
or to another scenario, the CarlaActorPool was added. All scenario
actors are added to this pool, and may be removed after completion of
the scenario. Alternatively, some may stay in the pool to be used after
the scenario exited.

Note:
Currently all scenario actors are removed after scenario completion

Change-Id: Ifd24cf41f22a66f2a1184b34216f79d1d7629f5d