Added InTriggerDistanceToLocationAlongRoute behavior
A new atomic behavior InTriggerDistanceToLocationAlongRoute was added.

As the name indiciates, the behavior calculates the distance to a given
location along a given route.

Therefore, a new method was added to tools/scenario_helper.py to
calculate the distance along a given route: get_distance_along_route()

The new behavior is used in the BasicScenario class to trigger a
scenario.

Change-Id: I91c9dd63334c49a6266536dfbe7b34fa562f9cff