Added watchdog to ScenarioManager and timeout handling for CARLA tick() (#475)
A watchdog is added to the ScenarioManager to detect and handle internal
timeouts as well as CARLA/Unreal crashes. Upon these a KeyboardInterrupt
is raised to trigger a clean shutdown of ScenarioRunner.

In addition, a timeout around CARLA tick() calls is implemented to avoid
getting stuck in a non-returning CARLA tick(). This is handled through
the newly created perform_carla_tick() in the CarlaDataProvider class.

Change-Id: If4cd17b7cf5f3e54fa93ea5e6e2026b0ba9ab13a