Update atomic_trigger_conditions.py
In the StandStill() class, change the EPSILON, 0.001, to the float value 0.1.

When a vehicle stops after cruising, even though it actually not moves at all, but the velocity could have a value larger than 0.001. 

The change makes StandStill() class work properly not only before vehicle's start but also after the vehicle's stop.
In the StandStill() class, change the EPSILON, 0.001, to the float value 0.1.

When a vehicle stops after cruising, even though it actually not moves at all, but the velocity could have a value larger than 0.001. 

The change makes StandStill() class work properly not only before vehicle's start but also after the vehicle's stop.